unsigned long -> gsize
git-svn-id: http://svn.irssi.org/repos/irssi/trunk@3132 dbcabf3a-b0e7-0310-adc4-f8d773084564

+++ src/core/net-nonblock.c
@@ -40,7 +40,7 @@ typedef struct {
 
 static int g_io_channel_write_block(GIOChannel *channel, void *data, int len)
 {
-        unsigned long ret;
+        gsize ret;
 	int err, sent;
 
 	sent = 0;
@@ -56,7 +56,7 @@ static int g_io_channel_write_block(GIOChannel *channel, void *data, int len)
 static int g_io_channel_read_block(GIOChannel *channel, void *data, int len)
 {
 	time_t maxwait;
-        unsigned long ret;
+        gsize ret;
 	int err, received;
 
 	maxwait = time(NULL)+2;
